test: share managed PCRE2 fixture across suites
