+++ .plans/pr650-managed-pcre2-test-fixtures-consensus.md
@@ -1,0 +1,22 @@
+# PR #650 managed-PCRE2 standalone test-fixture consensus
+
+- Specification: `.plans/pr650-managed-pcre2-test-fixtures.md`
+- Accepted SHA-256:
+  `d714d24092ce73bd3b97d7d153adca415fbea63a742f0b9b3a97641c74b73d45`
+- Review date: 2026-07-29
+- Review mechanism: Ollama
+
+## Final review round
+
+| Reviewer | Model | Verdict | Blockers |
+|---|---|---|---|
+| GLM 5.2 | `glm-5.2:cloud` | `ACCEPT` | None |
+| Kimi K2.7 Code | `kimi-k2.7-code:cloud` | `ACCEPT` | None |
+| MiniMax M3 | `minimax-m3:cloud` | `ACCEPT` | None |
+
+The first round produced one blocker from Kimi: a shared
+`tests/managed_pcre2_support.rs` file would be auto-discovered by Cargo as its
+own integration-test target. The specification was revised to use the
+non-auto-discovered `tests/support/managed_pcre2.rs` path and to harden target
+cache validation, sentinel selection, and diagnostic coverage. All three models
+reviewed that same revised digest and accepted it without blockers.
